Move Concat shape inference into concat.rs module
Relocate the Concat operator and its test from ops.rs into the
ops/concat.rs submodule, alongside Tile, to mirror the organization of
the corresponding ops in src/ops/.
